break up the metadata_file module into metadata_parser and metadata_file, and adjust all references
